requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
asyncpg
alembic
sqlalchemy
pydantic
pydantic-settings
pytest
bcrypt
python-jose[cryptography]